data-ml/requirements.txt
yfinance
pandas
numpy
matplotlib
optuna
seaborn
xgboost
scikit-learn
Flask==2.3.3
APScheduler==3.10.4
python-dotenv==1.0.1
